refactor: separate sort enum file
